requirements.txt
amqp==5.0.6
asgiref==3.3.1
autopep8==1.5.5
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
DAWG-Python==0.7.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-ckeditor==6.0.0
django-debug-toolbar==3.2
django-debug-toolbar-force==0.1.8
django-grappelli==2.14.3
django-heroku==0.3.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-nine==0.2.4
django-redis==4.12.1
django-timezone-field==4.1.2
djangorestframework==3.12.2
docopt==0.6.2
et-xmlfile==1.0.1
gunicorn==20.1.0
idna==2.10
kombu==5.0.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycodestyle==2.6.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
tablib==3.0.0
toml==0.10.2
Unidecode==1.2.0
urllib3==1.26.4
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
